tellspace-scaffold (1)/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
databases[postgresql]
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-dotenv
python-multipart
